+++ src/main/java/com/javack/ParaCasa/modelo/service/IPedidoService.java
@@ -7,8 +7,8 @@ import com.javack.ParaCasa.modelo.entity.Pedido;
 public interface IPedidoService {
 	public List<Pedido> listarTodos();
 	public void guardar(Pedido pedido);
-	public Pedido buscarPorId(int id);
-	public void eliminar (int id);
+	public Pedido buscarPorId(long id);
+	public void eliminar (long id);
 	
 
 }
